Add resolution-density adjustment to renderer
On my retina screen, WGo's current player looks fuzzy, which is a shame,
since it looks very nice otherwise. Luckily, there's an easy fix using
canvas: you can double the resolution by doubling the width and height
of the canvas elements, then setting the STYLE width and height
attributes to the correct width and height.

And it's easy to check the pixel density of the client: just check the
window.devicePixelRatio value.

This commit adds logic to increase the density of the renderer for any
pixel ratio.