Fix bug with board sizes that invoke setSize
I discovered looking at demo4.html that the 9x9 board was getting
doubled in size after my Retina fix. It turns out that non-19x19 boards
have `setDimensions()` called *after* the initial size setting, which
resulted in the size being multiplied undesirably.

This updates the `setDimensions` method to consult
`this.element.style.width` instead of `this.width` as the default, which
fixes the problem.